FDS_Source: radi: Parallel radiation spread
Introduces a second version of the radiation propagation.
This version removes the loop-carried dependencies of the original
version and therefore allows parallelisation with OpenMP.

The loop-carried dependencies basically cause a wave to spread
through the mesh, were the cells affected by the wave front are
dependent upon the cells in the slice lying behind them.

To remove the loop-carried dependencies the indices of these slices
are determined and stored in an array.
Subsequently the loop over the cells in each slice can be safely
parellelised with OpenMP.

To ease verification and code review this new version is only used
if USE_OPENMP is .true.


Former-commit-id: dc7d7c203b45572f7f53c06b4e00063916d87bf7